refactor(analysis): extract complexity debt helpers
